delete oklm

+++ controllers/delete_patients_ctrl.php
@@ -1,9 +1,9 @@
 <?php
-require_once dirname(__FILE__) . '/../model/Appointment.php';
 require_once dirname(__FILE__).'/../model/Patients.php';
+require_once dirname(__FILE__) . '/../model/Appointment.php';
 
 if (empty($_GET['id']) && empty($_POST['id'])) {
-    header('location : list_patients_ctrl.php');
+    header('location : liste_patients_ctrl.php');
     exit();
 }
 if (isset($_GET['id'])) {
@@ -17,9 +17,12 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $id = (int) $_POST['id'];
     $fullName = $_POST['fullName'];
     $patients = new Patients($id);
-    if ($patients->delete()) {
-        $deletePatientsSuccess = true;
-        header('refresh:5;list_patients_ctrl.php');
+    $appointment = new Appointment();
+    if ($appointment->deleteAppointment($id)) {
+        if ($patients->delete()) {
+            $deletePatientsSuccess = true;
+            header('refresh:1;liste_patients_ctrl.php');
+        }
     }
 }
 require_once dirname(__FILE__).'/../views/delete_patients.php';
